Fix persistence when delete student, pending fix persistence when delete course n teacher

+++ CourseService/src/main/java/com/Sophos/SophosUniversity/entities/Course.java
@@ -20,6 +20,6 @@ public class Course {
     private String course_name;
     private Integer amount_credits;
     private Integer available_places;
-    private Integer course_student_monitor_id;
+    private Long course_student_monitor_id;
 
 }
